requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
google-generativeai>=0.8.0
python-dotenv>=1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0
aiomysql>=0.2.0
firebase-admin>=6.0
openai>=1.0
